requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
PyJWT==2.10.1
setuptools==75.8.2
sqlparse==0.5.3
tzdata==2025.1
